user prompt for ssl generation
